zoop/backend/requirements/base.txt
# Django + DRF
Django>=5.0,<5.1
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4

# Auth
django-allauth>=0.63,<0.64

# Database
psycopg2-binary>=2.9,<3.0

# Background jobs
celery>=5.3,<5.4
django-celery-beat>=2.6.0,<2.10.0
redis>=5.0,<6.0

# Storage
boto3>=1.34,<1.35
django-storages>=1.14,<1.15

# Utilities
django-environ>=0.11,<0.12
python-dotenv>=1.0,<1.1

# API docs
drf-spectacular>=0.27,<0.28

# CORS
django-cors-headers>=4.3,<4.4

# SMS (Twilio)
twilio>=8.10,<8.11

# Email (Resend/SendGrid)
django-anymail>=10.2,<10.3

# Error tracking
sentry-sdk>=2.0,<2.1